Renaming the 'Iframe' folder to 'IFrame' to match the component type name case
