+++ src/layout/IFrame/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { IFrameComponent } from 'src/layout/Iframe/IFrameComponent';
+import { IFrameComponent } from 'src/layout/IFrame/IFrameComponent';
 import { PresentationComponent } from 'src/layout/LayoutComponent';
-import type { IFrameComponentProps } from 'src/layout/Iframe/IFrameComponent';
+import type { IFrameComponentProps } from 'src/layout/IFrame/IFrameComponent';
 
 export class IFrame extends PresentationComponent<'IFrame'> {
   render(props: IFrameComponentProps): JSX.Element | null {
